Model refactoring, Order Restaurants by votes count

+++ menuapp-web/src/main/java/org/turbodi/menuapp/web/dto/RestaurantDto.java
@@ -19,5 +19,5 @@ public class RestaurantDto {
 
     private boolean deleted;
 
-    private int votes;
+    private long votes;
 }
